+++ src/resources/af_move.h
@@ -19,8 +19,8 @@ typedef struct af_move_t {
     uint8_t block_stun;
     uint8_t throw_duration;
     uint8_t extra_string_selector;
-    float damage;
-    float raw_damage; // used to store the raw damage, unmodified by pilot stats
+    int16_t damage;
+    uint8_t raw_damage; // used to store the raw damage, unmodified by pilot stats
     har_endurance_t stun;
     str move_string;
     str footer_string;
